+++ packages/soda-nodejs/src/utils/spawnAsync.ts
@@ -16,16 +16,17 @@ export interface PromiseWithChildProcess<T> extends Promise<T> {
 }
 
 export type Options =
-    | SpawnOptionsWithoutStdio
-    | SpawnOptionsWithStdioTuple<StdioPipe, StdioPipe, StdioPipe>
-    | SpawnOptionsWithStdioTuple<StdioPipe, StdioPipe, StdioNull>
-    | SpawnOptionsWithStdioTuple<StdioPipe, StdioNull, StdioPipe>
-    | SpawnOptionsWithStdioTuple<StdioNull, StdioPipe, StdioPipe>
-    | SpawnOptionsWithStdioTuple<StdioPipe, StdioNull, StdioNull>
-    | SpawnOptionsWithStdioTuple<StdioNull, StdioPipe, StdioNull>
-    | SpawnOptionsWithStdioTuple<StdioNull, StdioNull, StdioPipe>
-    | SpawnOptionsWithStdioTuple<StdioNull, StdioNull, StdioNull>
-    | SpawnOptions
+
+        | SpawnOptionsWithoutStdio
+        | SpawnOptionsWithStdioTuple<StdioPipe, StdioPipe, StdioPipe>
+        | SpawnOptionsWithStdioTuple<StdioPipe, StdioPipe, StdioNull>
+        | SpawnOptionsWithStdioTuple<StdioPipe, StdioNull, StdioPipe>
+        | SpawnOptionsWithStdioTuple<StdioNull, StdioPipe, StdioPipe>
+        | SpawnOptionsWithStdioTuple<StdioPipe, StdioNull, StdioNull>
+        | SpawnOptionsWithStdioTuple<StdioNull, StdioPipe, StdioNull>
+        | SpawnOptionsWithStdioTuple<StdioNull, StdioNull, StdioPipe>
+        | SpawnOptionsWithStdioTuple<StdioNull, StdioNull, StdioNull>
+        | SpawnOptions
 
 let defaultOptions: Options = {}