+++ packages/sdnext/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-node
 
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.48.5
+
 ## 0.0.36
 
 ### Patch Changes